コンポーネントの命名を変更 理由：AppInnerだと役割がわからないから
